feat: add notification history feature with backend and frontend integration
- Implemented notification history retrieval in the backend with new controller and service.
- Added routes for fetching notifications and resending absent notifications.
- Created a new frontend page for displaying notification history with filtering options.
- Enhanced data models to include notification details and parent email for students.
- Updated attendance service to fetch notification history from the backend.
- Improved error handling and loading states in the notifications UI.
- Added new icons and styles for better user experience in the notifications section.

+++ backend/controllers/sessionController.js
@@ -194,9 +194,9 @@ exports.startSession = async (req, res) => {
 			if (Number.isFinite(homeClassId) && homeClassId > 0) {
 				await pool.query(
 					`INSERT INTO Enrollments (student_id, course_class_id)
-					 SELECT s.id, $1
+					 SELECT s.id, $1::int
 					 FROM Student s
-					 WHERE s.home_class_id = $2
+					 WHERE s.home_class_id = $2::int
 					   AND EXISTS (
 						   SELECT 1
 						   FROM Face_embeddings f
